~ Partial refactoring of tamarin_symbol_table builder
